Fix visual artifacting
We use a box shadow to do some sweet sweet animating between rep count colours,
however when combined with border radius you can see a small amount of the background peek through on the corners.
Super minor, but this commit just makes it so that the background matches the box shadow colour when the box shadow is present.
Adding a transition on the background colour is necessary because we need it to still be there for the entire lengtth of the box shadow animation. Hacky but works