Fix issue dismissing keyboard resetting value
On android, when the keyboard was dismissed, it would reset the value of the textfield

This commit changes the behaviour of the textfields to emit their values on input rather than onchange, which ensures the state is always up to date.

This was an issue before because on inputs which used decimals, it would strip the decimal point during the state loop, meaning the user couldn't input them.  I fixed
this by only updating the internal input's value from the props when it doesn't have focus