Fix iOS long press not firing
I was lazily using the oncontextmenu event to do long press, however this does not fire on iOS.

This commit switches to using a custom long press based on timers and pointer events (similar to what we do elsewhere).

As part of this I noticed that the dialog will be cancelled immediately on long press in iOS due to it firing some pointer events when the user hasn't moved enough.  To fix this, I just disabled scrim cancellation in those dialogs which are triggered by long press.

Fixes #263