Remember and reopen history and feed pages when navigating
The 'current session' has a behaviour where navigating with the nav bar remembers it and keeps it open, without having to go back to the 'base' page first.

This commit adds this behaviour for history and feed.  The advantage of this is now you can open a historical session, go back to your current session, and go back to the historical session, without having to find it in the list again.

Simply tap the nav button again to go to the 'base' page