Fix decimal separator
This commit amends the IOS app to introspect the decimal separator from the native IOS locale settings, and updates the culture to match.

The existing behaviour seems to be to use the language code only to set the decimal formatter, however ios allows users to configure number formatting independently of language/region.

Fixes #332