Implement aes enc/decryption
This is pretty gnarly. The packages I managed to get working remotely are pretty inconsistent, I've had to patch package them a bit to accept base64 encoding as the utf8 encoding kept dropping data for binary data. The RSA package is inconsistent between devices with the verify method param ordering