fix: Ensure changing a session date changes the set dates
Recently we moved to storing set times with dates. This means that changing the date in history did not update the stats or exports.

I have amended this code so that it will update the date of each set that is completed, while keeping it relative to the session's date.

This means that sessions which cross over midnight keep their set times consistent