fix: Typing in notes is janky
There are multiple reports of janky typing in the notes fields. This actually affects basically every field but it is most noticable in notes as the user types a lot. React native controlled inputs are pretty terrible I've noticed, so switching to defaultValue seems to work for these specific fields