Add weight unit as part of stored weights
This is the first part of the work to allow users to pick their unit of weight on a per weight basis, rather than globally.
This will be useful for certain exercises which use pounds rather than kilos or vice versa.

We have to still migrate all old weights to the preferred unit. I'm unsure if we should do this automatically, or use a wizard or something. Right now we support NIL for the unit, which stands for unspecified.