fix: Intermittent crash when selecting reps
Due to how we were tracking whether the potential set dialog is open, and sharing the state, it was possible for the app to accidentally send a setIndex of -1 in a race condition, which crashes the app when looking up which set to use.

This has been fixed by giving each potential set their own dialog for setting reps, and no longer tracking which one it is valid for with a set index. We never then use -1 as an index.