feat: Add a 'update plan' dialog to save changes made during a workout
It is common for a user to edit their workout as they complete it.
It is not entirely obvious that these changes do not apply to their plan.

This commit attempts to clarify this by showing a dialog at the end of their workout. This dialog shows the changes they made,
and allows them to select the ones they'd like to propagate back into their plan.

This will aim to replace the floating button that appears on the bottom left after changes are made

Remove update plan button