fix: Allow user to type timer values
There was a bug where the user would type their timer values for cardio exercises, but it would not persist. This was caused by the timer value being used over the typed value. This commit fixes the issue by ensuring that we always use the current duration value when displaying the timer.

Additionally, we fix an issue where if the user types a smaller denominated time unit (minutes) when there is a value in a higher unit (hours), it would clear the higher unit.

fixes: #624